compose: Extract compose_call module.
