+++ web/src/server_events_dispatch.js
@@ -9,6 +9,7 @@ import * as bot_data from "./bot_data";
 import * as browser_history from "./browser_history";
 import {buddy_list} from "./buddy_list";
 import * as compose from "./compose";
+import * as compose_call from "./compose_call";
 import * as compose_pm_pill from "./compose_pm_pill";
 import * as compose_recipient from "./compose_recipient";
 import * as composebox_typeahead from "./composebox_typeahead";
@@ -132,10 +133,10 @@ export function dispatch_normal_event(event) {
         case "has_zoom_token":
             page_params.has_zoom_token = event.value;
             if (event.value) {
-                for (const callback of compose.zoom_token_callbacks.values()) {
+                for (const callback of compose_call.zoom_token_callbacks.values()) {
                     callback();
                 }
-                compose.zoom_token_callbacks.clear();
+                compose_call.zoom_token_callbacks.clear();
             }
             break;
